+++ lib/custom_bloc.dart
@@ -11,6 +11,7 @@ export 'package:custom_bloc/src/base_bloc.dart';
 export 'package:custom_bloc/src/base_model.dart';
 export 'package:custom_bloc/src/custom_stream_bloc.dart';
 export 'package:custom_bloc/src/custom_bloc_builder.dart';
+export 'package:custom_bloc/src/custom_infinite_bloc_builder.dart';
 export 'package:custom_bloc/src/custom_infinite_scrolling_stream_bloc.dart';
 export 'package:custom_bloc/src/bloc_factories/base_bloc_infinite_loading_factory.dart';
 export 'package:custom_bloc/src/enum.dart';