+++ lib/src/base_bloc.dart
@@ -34,7 +34,10 @@ class BaseBloc<T, E> {
   }
 
   ///Stream declaration
-  var _behaviorSubject;
+  late Subject<BaseModel<T, E>> _behaviorSubject;
+
+  /// for checking if publish stream has value
+  BaseModel<T, E>? _latestPublishedValue;
 
   BaseBloc() {
     _behaviorSubject = _blocType == BlocType.alwaysReplayAllItemsToNewListener
@@ -42,36 +45,154 @@ class BaseBloc<T, E> {
         : _blocType == BlocType.broadcastOnlyIfIsCurrentlyListening
             ? PublishSubject<BaseModel<T, E>>()
             : BehaviorSubject<BaseModel<T, E>>.seeded(BaseModel<T, E>());
+
+    if (_behaviorSubject is PublishSubject) {
+      _behaviorSubject.listen((value) {
+        _latestPublishedValue = value;
+      });
+    }
   }
 
   ///getter for the stream
-  BehaviorSubject<BaseModel<T, E>> get behaviorSubject => _behaviorSubject;
+  Subject<BaseModel<T, E>> get behaviorSubject => _behaviorSubject;
 
   ///Getter for stream
-  ValueStream<BaseModel<T, E>> get stream => behaviorSubject.stream;
+  Stream<BaseModel<T, E>> get stream => behaviorSubject.stream;
 
   ///returns true if [bloc] is equal to [_baseModel.hasData] and [model] is not null
-  bool get blocHasData =>
-      behaviorSubject.hasValue &&
-      behaviorSubject.value.hasData &&
-      behaviorSubject.value.model != null;
+  bool get blocHasData {
+    if (_behaviorSubject is BehaviorSubject<BaseModel<T, E>>) {
+      final behaviorSubject =
+          _behaviorSubject as BehaviorSubject<BaseModel<T, E>>;
+      return behaviorSubject.hasValue &&
+          behaviorSubject.value.hasData &&
+          behaviorSubject.value.model != null;
+    } else if (_behaviorSubject is ReplaySubject<BaseModel<T, E>>) {
+      final replaySubject = _behaviorSubject as ReplaySubject<BaseModel<T, E>>;
 
-  ///returns true if [bloc] is equal to [_baseModel.hasError] and [error] is not null
-  bool get blocHasError =>
-      behaviorSubject.hasValue &&
-      behaviorSubject.value.hasError &&
-      behaviorSubject.value.error != null;
+      // final currentValue = replaySubject.valueOrNull;
+      // return currentValue != null && currentValue.hasData && currentValue.model != null;
+
+      bool hasEmittedData = false;
+      BaseModel<T, E>? latestValue;
+
+      final subscription = replaySubject.stream.listen((value) {
+        hasEmittedData = true;
+        latestValue = value;
+      });
+
+      subscription.cancel();
+
+      return hasEmittedData &&
+          latestValue != null &&
+          latestValue!.hasData &&
+          latestValue!.model != null;
+    } else if (_behaviorSubject is PublishSubject<BaseModel<T, E>>) {
+      return _latestPublishedValue != null &&
+          _latestPublishedValue!.hasData &&
+          _latestPublishedValue!.model != null;
+    } else {
+      return false;
+    }
+  }
+
+  /// [BlocType.alwaysReplayAllItemsToNewListener] would return true if there has ever
+  /// been any error added to the bloc
+  ///
+  /// [BlocType.alwaysBroadcastLatestToAllListeners] would always return true if only
+  /// the last item added to bloc was was error
+  ///
+  /// [BlocType.broadcastOnlyIfIsCurrentlyListening] would return true to listener only
+  /// if the last added item is an error after listener subscribe
+  bool get blocHasError {
+    if (_behaviorSubject is BehaviorSubject<BaseModel<T, E>>) {
+      final behaviorSubject =
+          _behaviorSubject as BehaviorSubject<BaseModel<T, E>>;
+      return behaviorSubject.hasValue &&
+          behaviorSubject.value.hasError &&
+          behaviorSubject.value.error != null;
+    } else if (_behaviorSubject is ReplaySubject<BaseModel<T, E>>) {
+      final replaySubject = _behaviorSubject as ReplaySubject<BaseModel<T, E>>;
+
+      bool hasEmittedData = false;
+      BaseModel<T, E>? latestValue;
+
+      final subscription = replaySubject.stream.listen((value) {
+        hasEmittedData = true;
+        latestValue = value;
+      });
+
+      subscription.cancel();
+
+      return hasEmittedData &&
+          latestValue != null &&
+          latestValue!.hasError &&
+          latestValue!.error != null;
+    } else if (_behaviorSubject is PublishSubject<BaseModel<T, E>>) {
+      return _latestPublishedValue != null &&
+          _latestPublishedValue!.hasError &&
+          _latestPublishedValue!.error != null;
+    } else {
+      return false;
+    }
+  }
 
   ///returns true if [bloc] is equal to [_baseModel.isLoading]
-  bool get blocIsLoading =>
-      behaviorSubject.hasValue && behaviorSubject.value.isLoading;
+  bool get blocIsLoading {
+    if (_behaviorSubject is BehaviorSubject<BaseModel<T, E>>) {
+      final behaviorSubject =
+          _behaviorSubject as BehaviorSubject<BaseModel<T, E>>;
+      return behaviorSubject.hasValue && behaviorSubject.value.isLoading;
+    } else if (_behaviorSubject is ReplaySubject<BaseModel<T, E>>) {
+      final replaySubject = _behaviorSubject as ReplaySubject<BaseModel<T, E>>;
+      bool hasEmittedData = false;
+      BaseModel<T, E>? latestValue;
 
-  ///returns true if [bloc] is equal to [_baseModel.isLoading]
-  bool get blocHasNoContent =>
-      behaviorSubject.hasValue && behaviorSubject.value.noContent;
+      final subscription = replaySubject.stream.listen((value) {
+        hasEmittedData = true;
+        latestValue = value;
+      });
+
+      subscription.cancel();
+
+      return hasEmittedData && latestValue != null && latestValue!.isLoading;
+    } else if (_behaviorSubject is PublishSubject<BaseModel<T, E>>) {
+      return _latestPublishedValue != null && _latestPublishedValue!.isLoading;
+    } else {
+      return false;
+    }
+  }
+
+  ///returns true if [bloc] is equal to [_baseModel.noContent]
+  bool get blocHasNoContent {
+    if (_behaviorSubject is BehaviorSubject<BaseModel<T, E>>) {
+      final behaviorSubject =
+          _behaviorSubject as BehaviorSubject<BaseModel<T, E>>;
+      return behaviorSubject.hasValue && behaviorSubject.value.noContent;
+    } else if (_behaviorSubject is ReplaySubject<BaseModel<T, E>>) {
+      // For ReplaySubject, access the latest value through a temporary subscription
+      final replaySubject = _behaviorSubject as ReplaySubject<BaseModel<T, E>>;
+      bool hasEmittedData = false;
+      BaseModel<T, E>? latestValue;
+
+      final subscription = replaySubject.stream.listen((value) {
+        hasEmittedData = true;
+        latestValue = value;
+      });
+
+      subscription.cancel();
+
+      return hasEmittedData && latestValue != null && latestValue!.noContent;
+    } else if (_behaviorSubject is PublishSubject<BaseModel<T, E>>) {
+      // For PublishSubject, use the manually tracked value
+      return _latestPublishedValue != null && _latestPublishedValue!.noContent;
+    } else {
+      return false;
+    }
+  }
 
   ///getter for nullable data
-  T? get dataNullable => blocHasData ? _behaviorSubject.value.model : null;
+  T? get dataNullable => blocHasData ? _getCurrentValue()?.model : null;
 
   ///getter for data
   ///Note: This will throw an exception, make sure to check if there is data by
@@ -81,11 +202,35 @@ class BaseBloc<T, E> {
       throw Exception(
           "Bloc does not appear to have any data. Make sure you have previously addDataToModel");
     }
-    return _behaviorSubject.value.model!;
+    return _getCurrentValue()!.model!;
   }
 
   ///getter for nullable error
-  E? get errorNullable => blocHasError ? _behaviorSubject.value.error : null;
+  E? get errorNullable => blocHasError ? _getCurrentValue()?.error : null;
+
+// Helper method to get current value from any Subject type
+  BaseModel<T, E>? _getCurrentValue() {
+    if (_behaviorSubject is BehaviorSubject<BaseModel<T, E>>) {
+      return (_behaviorSubject as BehaviorSubject<BaseModel<T, E>>).value;
+    } else if (_behaviorSubject is ReplaySubject<BaseModel<T, E>>) {
+      // For ReplaySubject, we need to use a subscription to get the current value
+      final replaySubject = _behaviorSubject as ReplaySubject<BaseModel<T, E>>;
+      BaseModel<T, E>? latestValue;
+
+      final subscription = replaySubject.stream.listen((value) {
+        latestValue = value;
+      });
+
+      subscription.cancel();
+
+      return latestValue;
+    } else if (_behaviorSubject is PublishSubject<BaseModel<T, E>>) {
+      // For PublishSubject, use our manually tracked value
+      return _latestPublishedValue;
+    }
+
+    return null;
+  }
 
   ///getter for error.
   ///Note: This will throw an exception, make sure to check if there is error by
@@ -95,7 +240,7 @@ class BaseBloc<T, E> {
       throw Exception(
           "Bloc does not appear to have any error. Make sure you have previously addToError");
     }
-    return _behaviorSubject.value.error!;
+    return _getCurrentValue()!.error!;
   }
 
   ///Getter for Base object